// src/modules/users/types/userStream.types.ts
interface UserStreamData {
  company_id: string;
  created_at: number;
  is_register: boolean;
  company_role_id: string;
  pin: string;
  updated_at: number;
  user_id: string;
  phone: string;
  identity_card_type: string;
  name: string;
  limit: number;
  id: string;
  email: string;
  table: string;
  username: string;
  status: number;
  identity_card_number: string;
}

export default UserStreamData;